Adding the basics for the starting settings of the game. The 1942SpringStartingConfig.json encapsulates the data on the starting cards.
Obviously this is in JSON format.

This configs will be used to populate WorldPower objects.
Faction is an enum (AXIS or ALLIES)

I used http://jsonlint.com/ to make sure the .json file was valid.

The .jsonstub was used to fill out each WorldPower quicker and may
be used again if we are using the game for more players (9?)